==> legacy/1_1_data_processing_growth_.ipynb <==
import os

import numpy as np
import pandas as pd

# SEM length & diameter

# 2020 Summer

DIRECTORY = './data/2020_S/destructive/'
file_list = os.listdir(DIRECTORY)
dataset_list = [file for file in file_list if file.endswith('.xlsx')]
dataset_list.sort()

dates = ['0324', '0421', '0522', '0617', '0707', '0708', '0709']

node_len_df = {}
node_dia_df = {}
_ = 0
for FILENAME in dataset_list[-1:]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=1)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        node_len_df[key] = np.append(temp_df[key]['stem length(cm)'].dropna().iloc[-2:].values, temp_df[key]['stem length(cm).1'].dropna().iloc[-2:].values).astype('float').mean()
        node_dia_df[key] = np.append(temp_df[key]['stem diameter(mm)'].dropna().iloc[-2:].values, temp_df[key]['stem diameter(mm).1'].dropna().iloc[-2:].values).astype('float').mean()
    _ += 1 
node_len_df = pd.DataFrame(node_len_df, index=['node_length']).T
node_dia_df = pd.DataFrame(node_dia_df, index=['node_diameter']).T

node_len = []
node_dia = []
for _ in range(4):
    node_len.append(node_len_df.filter(like=dates[_], axis=0).values.T.squeeze())
    node_dia.append(node_dia_df.filter(like=dates[_], axis=0).values.T.squeeze())

node_len_df = {}
node_dia_df = {}
_ = 0
for FILENAME in dataset_list[:-1]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=1)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
        try:
            node_len_df[key + '_' + dates[-3:][_]] = np.append(temp_df[key]['internode length'].dropna().iloc[-2:].values, temp_df[key]['internode length.1'].dropna().iloc[-2:].values).astype('float').mean()
            node_dia_df[key + '_' + dates[-3:][_]] = np.append(temp_df[key]['internode diameter'].dropna().iloc[-2:].values, temp_df[key]['Unnamed: 12'].dropna().iloc[-2:].values).astype('float').mean()
        except KeyError:
            node_len_df[key + '_' + dates[-3:][_]] = np.append(temp_df[key]['internode length'].dropna().iloc[-2:].values, temp_df[key]['Unnamed: 10'].dropna().iloc[-2:].values).astype('float').mean()
            node_dia_df[key + '_' + dates[-3:][_]] = np.append(temp_df[key]['internode diameter'].dropna().iloc[-2:].values, temp_df[key]['Unnamed: 12'].dropna().iloc[-2:].values).astype('float').mean()
            
    _ += 1 
node_len_df = pd.DataFrame(node_len_df, index=['node_length']).T
node_dia_df = pd.DataFrame(node_dia_df, index=['node_diameter']).T

a2_len_df = node_len_df.filter(like='A2', axis=0)
a2_dia_df = node_dia_df.filter(like='A2', axis=0)

for _ in range(3):
    node_len.append(a2_len_df.filter(like=dates[-3:][_], axis=0).values.T.squeeze())
    node_dia.append(a2_dia_df.filter(like=dates[-3:][_], axis=0).values.T.squeeze())

node_len_df = pd.DataFrame(node_len, index=['2020-03-24', '2020-04-21', '2020-05-22', '2020-06-17', '2020-07-07', '2020-07-08', '2020-07-09'])
node_dia_df = pd.DataFrame(node_dia, index=['2020-03-24', '2020-04-21', '2020-05-22', '2020-06-17', '2020-07-07', '2020-07-08', '2020-07-09'])

node_len_df.index = pd.DatetimeIndex(node_len_df.index)
node_dia_df.index = pd.DatetimeIndex(node_dia_df.index)

node_len_df.to_csv('./results/2020_S/node_length.csv')
node_dia_df.to_csv('./results/2020_S/node_diameter.csv')

# 2020 Winter

DIRECTORY = './data/2020_W/destructive/'
file_list = os.listdir(DIRECTORY)
dataset_list = [file for file in file_list if file.endswith('.xlsx')]
dataset_list.sort()

dates = ['2020-09-09', '2020-10-14', '2020-11-13', '2020-12-11']

node_len_df = {}
node_dia_df = {}
_ = 0
for FILENAME in dataset_list[:-1]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=1)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        node_len_df[dates[_] + '_' + key] = np.append(temp_df[key]['internode length'].dropna().iloc[-2:].values, temp_df[key]['internode length.1'].dropna().iloc[-2:].values).astype('float').mean()
        node_dia_df[dates[_] + '_' + key] = np.append(temp_df[key]['internode diameter'].dropna().iloc[-2:].values, temp_df[key]['internode diameter.1'].dropna().iloc[-2:].values).astype('float').mean()
    _ += 1 
node_len_df = pd.DataFrame(node_len_df, index=['node_length']).T
node_dia_df = pd.DataFrame(node_dia_df, index=['node_diameter']).T

node_len_df = node_len_df.drop(node_len_df.filter(like='RB', axis=0).index)
node_dia_df = node_dia_df.drop(node_dia_df.filter(like='RB', axis=0).index)

node_len = []
node_dia = []
for _ in range(4):
    node_len.append(node_len_df.filter(like=dates[_], axis=0).values.T.squeeze())
    node_dia.append(node_dia_df.filter(like=dates[_], axis=0).values.T.squeeze())

node_len_df = {}
node_dia_df = {}
_ = 0
for FILENAME in dataset_list[-1:]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=0)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
        node_len_df[key] = np.append(temp_df[key]['Main stem'].dropna().iloc[-2:].values, temp_df[key]['Side stem'].dropna().iloc[-2:].values).astype('float').mean()
        node_dia_df[key] = np.append(temp_df[key]['Main stem.1'].dropna().iloc[-2:].values, temp_df[key]['Side stem '].dropna().iloc[-2:].values).astype('float').mean()
            
    _ += 1 
node_len_df = pd.DataFrame(node_len_df, index=['node_length']).T
node_dia_df = pd.DataFrame(node_dia_df, index=['node_diameter']).T

a2_len_df = node_len_df.filter(like='a2', axis=0)
a2_dia_df = node_dia_df.filter(like='a2', axis=0)

node_len.append(a2_len_df.values.T.squeeze())
node_dia.append(a2_dia_df.values.T.squeeze())

node_len_df = pd.DataFrame(node_len, index=['2020-09-09', '2020-10-14', '2020-11-13', '2020-12-11', '2021-01-25'])
node_dia_df = pd.DataFrame(node_dia, index=['2020-09-09', '2020-10-14', '2020-11-13', '2020-12-11', '2021-01-25'])

node_len_df.index = pd.DatetimeIndex(node_len_df.index)
node_dia_df.index = pd.DatetimeIndex(node_dia_df.index)

node_len_df.loc['2021-01-25'] /= 10

node_len_df.to_csv('./results/2020_W/node_length.csv')
node_dia_df.to_csv('./results/2020_W/node_diameter.csv')

# Leaf area

# 2020 Summer

DIRECTORY = './data/2020_S/destructive/'
file_list = os.listdir(DIRECTORY)
dataset_list = [file for file in file_list if file.endswith('.xlsx')]
dataset_list.sort()

dates = ['0324', '0421', '0522', '0617', '0707', '0708', '0709']

leaf_area_df = {}
_ = 0
for FILENAME in dataset_list[-1:]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=0)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        leaf_area_df[key] = [temp_df[key]['leaf area'][0]]
    _ += 1 
leaf_area_df = pd.DataFrame.from_dict(leaf_area_df).T
leaf_area_df.columns = ['leaf area']

leaf_areas = []
for _ in range(4):
    leaf_areas.append(leaf_area_df.filter(like=dates[_], axis=0).values.T.squeeze())

leaf_area_df = {}
_ = 0
for FILENAME in dataset_list[:-1]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=0)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        leaf_area_df[key + '_' + dates[-3:][_]] = [temp_df[key]['leaf area'][0]]
    _ += 1 
leaf_area_df = pd.DataFrame.from_dict(leaf_area_df).T
leaf_area_df.columns = ['leaf area']

a2_df = leaf_area_df.filter(like='A2', axis=0)

for _ in range(3):
    leaf_areas.append(a2_df.filter(like=dates[-3:][_], axis=0).values.T.squeeze())

leaf_area_df = pd.DataFrame(leaf_areas, index=['2020-03-24', '2020-04-21', '2020-05-22', '2020-06-17', '2020-07-07', '2020-07-08', '2020-07-09'])

leaf_area_df.index = pd.DatetimeIndex(leaf_area_df.index)

leaf_area_df.to_csv('./results/2020_S/leaf_area.csv')

# 2020 Winter

DIRECTORY = './data/2020_W/destructive/'
file_list = os.listdir(DIRECTORY)
dataset_list = [file for file in file_list if file.endswith('.xlsx')]
dataset_list.sort()

dates = ['2020-09-09', '2020-10-14', '2020-11-13', '2020-12-11', '2021-01-25']

leaf_area_df = {}
_ = 0
for FILENAME in dataset_list:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=0)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        leaf_area_df[dates[_] + '_' + key] = [temp_df[key].loc[:2, 'leaf area'].sum()]
    _ += 1 
leaf_area_df = pd.DataFrame.from_dict(leaf_area_df).T
leaf_area_df.columns = ['leaf_area']

leaf_area_df = leaf_area_df.drop(leaf_area_df.filter(like='RB', axis=0).index)
leaf_area_df = leaf_area_df.drop(leaf_area_df.filter(like='a3', axis=0).index)

leaf_area_df.index = ['2020-09-09']*3 + ['2020-10-14']*6 + ['2020-11-13']*6 + ['2020-12-11']*3 + ['2021-01-25']*15

leaf_area_df.index = pd.DatetimeIndex(leaf_area_df.index)

leaf_area_df.to_csv('./results/2020_W/leaf_area.csv')

# Plant heights & num nodes

# 2020 Summer

DIRECTORY = './data/2020_S/destructive/'
file_list = os.listdir(DIRECTORY)
dataset_list = [file for file in file_list if file.endswith('.xlsx')]
dataset_list.sort()

dates = ['0324', '0421', '0522', '0617', '0707', '0708', '0709']

plant_height_df = {}
num_nodes_df = {}
for FILENAME in dataset_list[-1:]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=1)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
        temp_target = temp_df[key].loc[:, 'stem length(cm)':'stem diameter(mm).1'].dropna(axis=0, how='all')
        temp_target.loc[:, 'stem length(cm).1'] = temp_target['stem length(cm).1'].combine_first(temp_target['stem length(cm)'])
        temp_target.loc[:, 'stem diameter(mm).1'] = temp_target['stem diameter(mm).1'].combine_first(temp_target['stem diameter(mm)'])
        
        plant_height_df[key] = max(temp_target.iloc[:, :2].sum())
        num_nodes_df[key] = temp_target.shape[0]
plant_height_df = pd.DataFrame.from_dict([plant_height_df]).T
plant_height_df.columns = ['plant_height']
num_nodes_df = pd.DataFrame.from_dict([num_nodes_df]).T
num_nodes_df.columns = ['num_nodes_df']

plant_height = []
for _ in range(4):
    plant_height.append(plant_height_df.filter(like=dates[_], axis=0).values.T.squeeze())
num_nodes = []
for _ in range(4):
    num_nodes.append(num_nodes_df.filter(like=dates[_], axis=0).values.T.squeeze())

plant_height_df = {}
num_nodes_df = {}
_ = 0
for FILENAME in dataset_list[:-1]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=1)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
        temp_target = temp_df[key].loc[:, 'internode length':'Unnamed: 12'].dropna(axis=0, how='all')
        try:
            temp_target.loc[:, 'internode length.1'] = temp_target['internode length.1'].combine_first(temp_target['internode length'])
        except KeyError:
            temp_target.loc[:, 'Unnamed: 10'] = temp_target['Unnamed: 10'].combine_first(temp_target['internode length'])
        temp_target.loc[:, 'Unnamed: 12'] = temp_target['Unnamed: 12'].combine_first(temp_target['internode diameter'])
        
        plant_height_df[key + '_' + dates[-3:][_]] = max(temp_target.iloc[:, :2].sum())
        num_nodes_df[key + '_' + dates[-3:][_]] = temp_target.shape[0]
    _ += 1
plant_height_df = pd.DataFrame.from_dict([plant_height_df]).T
plant_height_df.columns = ['plant_height']
num_nodes_df = pd.DataFrame.from_dict([num_nodes_df]).T
num_nodes_df.columns = ['num_nodes_df']

a2_df = plant_height_df.filter(like='A2', axis=0)
a2_df_ = num_nodes_df.filter(like='A2', axis=0)

for _ in range(3):
    plant_height.append(a2_df.filter(like=dates[-3:][_], axis=0).values.T.squeeze())
    num_nodes.append(a2_df_.filter(like=dates[-3:][_], axis=0).values.T.squeeze())

plant_height_df = pd.DataFrame(plant_height, index=['2020-03-24', '2020-04-21', '2020-05-22', '2020-06-17', '2020-07-07', '2020-07-08', '2020-07-09'])
num_nodes_df = pd.DataFrame(num_nodes, index=['2020-03-24', '2020-04-21', '2020-05-22', '2020-06-17', '2020-07-07', '2020-07-08', '2020-07-09'])

plant_height_df.index = pd.DatetimeIndex(plant_height_df.index)
num_nodes_df.index = pd.DatetimeIndex(num_nodes_df.index)

plant_height_df.to_csv('./results/2020_S/plant_height.csv')
num_nodes_df.to_csv('./results/2020_S/num_nodes.csv')

# 2020 Winter

DIRECTORY = './data/2020_W/destructive/'
file_list = os.listdir(DIRECTORY)
dataset_list = [file for file in file_list if file.endswith('.xlsx')]
dataset_list.sort()

dates = ['2020-09-09', '2020-10-14', '2020-11-13', '2020-12-11', '2021-01-25']

# plant_height_df = {}
# num_nodes_df = {}
# for FILENAME in dataset_list:
#     temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=0)
#     for key in temp_df.keys():
#         if temp_df[key].shape[0] == 0:
#             continue
        
#         plant_height_df[key] = temp_df[key].loc[0, 'Plant height']
# _ = pd.DataFrame.from_dict([plant_height_df])
# _.columns = plant_height_df.keys()
# plant_height_df = _.T
# plant_height_df.columns = ['plant_height']

plant_height_df = {}
num_nodes_df = {}
_ = 0
for FILENAME in dataset_list:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=1)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
        try:
            temp_target = temp_df[key].loc[:, 'internode length':'internode diameter.1'].dropna(axis=0, how='all')
            temp_target.loc[:, 'internode length.1'] = temp_target['internode length.1'].combine_first(temp_target['internode length'])
            temp_target.loc[:, 'internode diameter.1'] = temp_target['internode diameter.1'].combine_first(temp_target['internode diameter'])
        except KeyError:
            temp_target = temp_df[key].loc[:, 'stem length(mm)':'stem diameter(mm).1'].dropna(axis=0, how='all')
            temp_target.loc[:, 'stem length(cm)'] = temp_target['stem length(cm)'].combine_first(temp_target['stem length(mm)'])
            temp_target.loc[:, 'stem diameter(mm).1'] = temp_target['stem diameter(mm).1'].combine_first(temp_target['stem diameter(mm)'])
        
        plant_height_df[dates[_] + '_' + key] = max(temp_target.iloc[:, :2].sum())
        num_nodes_df[dates[_] + '_' + key] = temp_target.shape[0]
    _ += 1
plant_height_df = pd.DataFrame.from_dict([plant_height_df]).T
plant_height_df.columns = ['plant_height']
num_nodes_df = pd.DataFrame.from_dict([num_nodes_df]).T
num_nodes_df.columns = ['num_nodes_df']

plant_height_df = plant_height_df.drop(plant_height_df.filter(like='RB', axis=0).index)
plant_height_df = plant_height_df.drop(plant_height_df.filter(like='a3', axis=0).index)
num_nodes_df = num_nodes_df.drop(num_nodes_df.filter(like='RB', axis=0).index)
num_nodes_df = num_nodes_df.drop(num_nodes_df.filter(like='a3', axis=0).index)

plant_height_df.index = ['2020-09-09']*3 + ['2020-10-14']*6 + ['2020-11-13']*6 + ['2020-12-11']*3 + ['2021-01-25']*15
num_nodes_df.index = ['2020-09-09']*3 + ['2020-10-14']*6 + ['2020-11-13']*6 + ['2020-12-11']*3 + ['2021-01-25']*15

plant_height = []
num_nodes = []
for _ in range(5):
    plant_height.append(plant_height_df.filter(like=plant_height_df.index.unique()[_], axis=0).values.T.squeeze())
    num_nodes.append(num_nodes_df.filter(like=num_nodes_df.index.unique()[_], axis=0).values.T.squeeze())

plant_height_df = pd.DataFrame(plant_height, index=['2020-09-09', '2020-10-14', '2020-11-13', '2020-12-11', '2021-01-25'])
num_nodes_df = pd.DataFrame(num_nodes, index=['2020-09-09', '2020-10-14', '2020-11-13', '2020-12-11', '2021-01-25'])

plant_height_df.index = pd.DatetimeIndex(plant_height_df.index)
num_nodes_df.index = pd.DatetimeIndex(num_nodes_df.index)

plant_height_df.loc['2021-01-25'] /= 10

plant_height_df.to_csv('./results/2020_W/plant_height.csv')
num_nodes_df.to_csv('./results/2020_W/num_nodes.csv')

# Weights

# 2020 Summer

# Control

DIRECTORY = './data/2020_S/destructive/'
file_list = os.listdir(DIRECTORY)
dataset_list = [file for file in file_list if file.endswith('.xlsx')]
dataset_list.sort()

dataset_list

FW_df = {}
_ = 0
for FILENAME in dataset_list[-1:]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=0)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        FW_df[key] = temp_df[key][['Stem FW', 'Leaf FW', 'petiole FW', 'Fruit count', 'Idv fruit FW', 'Plant height']].iloc[:10, :].sum()
    _ += 1 
FW_df = pd.DataFrame.from_dict(FW_df).T

FW_df.index = ['2020-03-24']*3 + ['2020-04-21']*3 + ['2020-05-22']*3 + ['2020-06-17']*4
__ = FW_df

FW_df = {}
_ = 0
for FILENAME in dataset_list[:-1]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=0)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        FW_df[key] = temp_df[key][['Stem FW', 'Leaf FW', 'petiole FW', 'Fruit count', 'Idv fruit FW', 'Plant height']].iloc[:10, :].sum()
    _ += 1 
FW_df = pd.DataFrame.from_dict(FW_df).T

FW_df = FW_df.filter(like='A2', axis=0)
FW_df.index = ['2020-07-07' for i in range(len(FW_df.index))]

FW_df = pd.concat([__, FW_df], axis=0, sort=False)

### Dry weights

DW_df = {}
_ = 0
for FILENAME in dataset_list[-1:]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=22)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        DW_df[key] = temp_df[key][['Stem DW', 'Leaf DW', 'petiole DW', 'Fruit DW']].iloc[:3, :].sum()
    _ += 1 
DW_df = pd.DataFrame.from_dict(DW_df).T

DW_df.index = ['2020-03-24']*3 + ['2020-04-21']*3 + ['2020-05-22']*3 + ['2020-06-17']*4
__ = DW_df

DW_df = {}
_ = 0
for FILENAME in dataset_list[:-1]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=22)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        DW_df[key] = temp_df[key][['Stem DW', 'Leaf DW', 'petiole DW', 'Fruit DW']].iloc[:3, :].sum()
    _ += 1 
DW_df = pd.DataFrame.from_dict(DW_df).T

DW_df = DW_df.filter(like='A2', axis=0)
DW_df.index = ['2020-07-07' for i in range(len(DW_df.index))]

DW_df = pd.concat([__, DW_df], axis=0, sort=False)

weight_df = pd.concat([FW_df, DW_df], axis=1, sort=False)

weight_df.to_csv('./results/weight_ct.csv')

# 2020 Winter

# Far-red

DIRECTORY = './data/2020_S/destructive/'
file_list = os.listdir(DIRECTORY)
dataset_list = [file for file in file_list if file.endswith('.xlsx')]
dataset_list.sort()

dataset_list

FW_df = {}
_ = 0
for FILENAME in dataset_list[:-1]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=0)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        FW_df[key] = temp_df[key][['Stem FW', 'Leaf FW', 'petiole FW', 'Fruit count', 'Idv fruit FW', 'Plant height']].iloc[:10, :].sum()
    _ += 1 
FW_df = pd.DataFrame.from_dict(FW_df).T

FW_df = FW_df.filter(like='A3', axis=0)
FW_df.index = ['2020-07-07' for i in range(len(FW_df.index))]

### Dry weights

DW_df = {}
_ = 0
for FILENAME in dataset_list[:-1]:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=22)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        DW_df[key] = temp_df[key][['Stem DW', 'Leaf DW', 'petiole DW', 'Fruit DW']].iloc[:3, :].sum()
    _ += 1 
DW_df = pd.DataFrame.from_dict(DW_df).T

DW_df = DW_df.filter(like='A3', axis=0)
DW_df.index = ['2020-07-07' for i in range(len(DW_df.index))]

weight_df = pd.concat([FW_df, DW_df], axis=1, sort=False)

weight_df.to_csv('./results/2020_S/weight_fr.csv')

# interlighting

# 2020 Winter Fresh weights

DIRECTORY = './data/2020_W/destructive/'
file_list = os.listdir(DIRECTORY)
dataset_list = [file for file in file_list if file.endswith('.xlsx')]
dataset_list.sort()

dataset_list

FW_df = {}
_ = 0
for FILENAME in dataset_list:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=0)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        FW_df[key] = temp_df[key][['Stem FW', 'Leaf FW', 'petiole FW', 'Fruit count', 'Idv fruit FW', 'Plant height']].iloc[:10, :].sum()
    _ += 1 
FW_df = pd.DataFrame.from_dict(FW_df).T

FW_df.index = ['2020-09-09']*3 + ['2020-10-14']*6 + ['2020-11-13']*6 + ['2020-12-11_']*3 + ['2020-12-11']*3
__ = FW_df

__rb = __.loc['2020-12-11_']
__rb.index = ['2020-12-11']*3
__ = __.drop('2020-12-11_')

### Dry weights

DW_df = {}
_ = 0
for FILENAME in dataset_list:
    temp_df = pd.read_excel(DIRECTORY + FILENAME, sheet_name=None, skiprows=22)
    for key in temp_df.keys():
        if temp_df[key].shape[0] == 0:
            continue
            
        DW_df[key] = temp_df[key][['Stem DW', 'Leaf DW', 'petiole DW', 'Fruit DW']].iloc[:3, :].sum()
    _ += 1 
DW_df = pd.DataFrame.from_dict(DW_df).T

DW_df.index = ['2020-09-09']*3 + ['2020-10-14']*6 + ['2020-11-13']*6 + ['2020-12-11_']*3 + ['2020-12-11']*3

DW_rb_df = DW_df.loc['2020-12-11_']
DW_rb_df.index = ['2020-12-11']*3
DW_df = DW_df.drop('2020-12-11_')

weight_ct_df = pd.concat([__, DW_df], axis=1, sort=False)
weight_rb_df = pd.concat([__rb, DW_rb_df], axis=1, sort=False)

weight_ct_df.to_csv('./results/2020_W/ct_weight.csv')
weight_rb_df.to_csv('./results/2020_W/rb_weight.csv')
