==> legacy/2_GAIN-interpolation-halted.ipynb <==
## Package loading

import tensorflow as tf
from tensorflow.keras import *
import numpy as np
import pandas as pd
from tqdm import tqdm

from GAIN.utils import normalization, renormalization, rounding
from GAIN.utils import xavier_init
from GAIN.utils import binary_sampler, uniform_sampler, sample_batch_index

# def gain (data_x, gain_parameters):
#     '''Impute missing values in data_x

#     Args:
#     - data_x: original data with missing values
#     - gain_parameters: GAIN network parameters:
#       - batch_size: Batch size
#       - hint_rate: Hint rate
#       - alpha: Hyperparameter
#       - iterations: Iterations

#     Returns:
#     - imputed_data: imputed data
#     '''
#     # Define mask matrix
#     data_m = 1-np.isnan(data_x)

#     # System parameters
#     batch_size = 128
#     hint_rate = 0.9
#     alpha = 100
#     iterations = 10000

#     # Other parameters
#     no, dim = data_x.shape

#     # Hidden state dimensions
#     h_dim = int(dim)

#     # Normalization
#     norm_data, norm_parameters = normalization(data_x)
#     norm_data_x = np.nan_to_num(norm_data, 0)

#     ## GAIN architecture   
#     # Input placeholders
#     # Data vector
#     X = tf.placeholder(tf.float32, shape = [None, dim])
#     # Mask vector 
#     M = tf.placeholder(tf.float32, shape = [None, dim])
#     # Hint vector
#     H = tf.placeholder(tf.float32, shape = [None, dim])

#     # Discriminator variables
#     D_W1 = tf.Variable(xavier_init([dim*2, h_dim])) # Data + Hint as inputs
#     D_b1 = tf.Variable(tf.zeros(shape = [h_dim]))

#     D_W2 = tf.Variable(xavier_init([h_dim, h_dim]))
#     D_b2 = tf.Variable(tf.zeros(shape = [h_dim]))

#     D_W3 = tf.Variable(xavier_init([h_dim, dim]))
#     D_b3 = tf.Variable(tf.zeros(shape = [dim]))  # Multi-variate outputs

#     theta_D = [D_W1, D_W2, D_W3, D_b1, D_b2, D_b3]

#     #Generator variables
#     # Data + Mask as inputs (Random noise is in missing components)
#     G_W1 = tf.Variable(xavier_init([dim*2, h_dim]))  
#     G_b1 = tf.Variable(tf.zeros(shape = [h_dim]))

#     G_W2 = tf.Variable(xavier_init([h_dim, h_dim]))
#     G_b2 = tf.Variable(tf.zeros(shape = [h_dim]))

#     G_W3 = tf.Variable(xavier_init([h_dim, dim]))
#     G_b3 = tf.Variable(tf.zeros(shape = [dim]))

#     theta_G = [G_W1, G_W2, G_W3, G_b1, G_b2, G_b3]
  
#     ## GAIN functions
#     # Generator
#     def generator(x,m):
#         # Concatenate Mask and Data
#         inputs = tf.concat(values = [x, m], axis = 1) 
#         G_h1 = tf.nn.relu(tf.matmul(inputs, G_W1) + G_b1)
#         G_h2 = tf.nn.relu(tf.matmul(G_h1, G_W2) + G_b2)   
#         # MinMax normalized output
#         G_prob = tf.nn.sigmoid(tf.matmul(G_h2, G_W3) + G_b3) 
#         return G_prob
      
#     # Discriminator
#     def discriminator(x, h):
#         # Concatenate Data and Hint
#         inputs = tf.concat(values = [x, h], axis = 1) 
#         D_h1 = tf.nn.relu(tf.matmul(inputs, D_W1) + D_b1)  
#         D_h2 = tf.nn.relu(tf.matmul(D_h1, D_W2) + D_b2)
#         D_logit = tf.matmul(D_h2, D_W3) + D_b3
#         D_prob = tf.nn.sigmoid(D_logit)
#         return D_prob
  
#     ## GAIN structure
#     # Generator
#     G_sample = generator(X, M)

#     # Combine with observed data
#     Hat_X = X * M + G_sample * (1-M)

#     # Discriminator
#     D_prob = discriminator(Hat_X, H)

#     ## GAIN loss
#     D_loss_temp = -tf.reduce_mean(M * tf.log(D_prob + 1e-8) \
#                                 + (1-M) * tf.log(1. - D_prob + 1e-8)) 

#     G_loss_temp = -tf.reduce_mean((1-M) * tf.log(D_prob + 1e-8))

#     MSE_loss = \
#     tf.reduce_mean((M * X - M * G_sample)**2) / tf.reduce_mean(M)

#     D_loss = D_loss_temp
#     G_loss = G_loss_temp + alpha * MSE_loss 

#     ## GAIN solver
#     D_solver = tf.train.AdamOptimizer().minimize(D_loss, var_list=theta_D)
#     G_solver = tf.train.AdamOptimizer().minimize(G_loss, var_list=theta_G)

#     ## Iterations
#     sess = tf.Session()
#     sess.run(tf.global_variables_initializer())
   
#     # Start Iterations
#     for it in tqdm(range(iterations)):    
      
#         # Sample batch
#         batch_idx = sample_batch_index(no, batch_size)
#         X_mb = norm_data_x[batch_idx, :]  
#         M_mb = data_m[batch_idx, :]  
#         # Sample random vectors  
#         Z_mb = uniform_sampler(0, 0.01, batch_size, dim) 
#         # Sample hint vectors
#         H_mb_temp = binary_sampler(hint_rate, batch_size, dim)
#         H_mb = M_mb * H_mb_temp

#         # Combine random vectors with observed vectors
#         X_mb = M_mb * X_mb + (1-M_mb) * Z_mb 

#         _, D_loss_curr = sess.run([D_solver, D_loss_temp], 
#                                   feed_dict = {M: M_mb, X: X_mb, H: H_mb})
#         _, G_loss_curr, MSE_loss_curr = \
#         sess.run([G_solver, G_loss_temp, MSE_loss],
#                  feed_dict = {X: X_mb, M: M_mb, H: H_mb})
            
#     ## Return imputed data      
#     Z_mb = uniform_sampler(0, 0.01, no, dim) 
#     M_mb = data_m
#     X_mb = norm_data_x          
#     X_mb = M_mb * X_mb + (1-M_mb) * Z_mb 

#     imputed_data = sess.run([G_sample], feed_dict = {X: X_mb, M: M_mb})[0]

#     imputed_data = data_m * norm_data_x + (1-data_m) * imputed_data

#     # Renormalization
#     imputed_data = renormalization(imputed_data, norm_parameters)  

#     # Rounding
#     imputed_data = rounding(imputed_data, data_x)  

#     return imputed_data

## Parameters

BATCH_SIZE = 128
HINT_RATE = 0.9
ALPHA = 100
ITERATIONS = 10000

## Data loading

data_df = pd.read_csv('./results/SW2_greenhouse.csv', index_col='Unnamed: 0')
data_df = data_df.astype('float32')
data_x = data_df.values
data_m = 1-np.isnan(data_x) # mask
num_data, dim = data_x.shape
hidden_dim = int(dim)

## Normalization

norm_data, norm_parameters = normalization(data_x)
norm_data_x = np.nan_to_num(norm_data, 0)

class Generator(Model):
    def __init__(self, num_nodes, output_size):
        super(Generator, self).__init__()
        self.n = num_nodes
        self.o = output_size
                
        self.dense1 = layers.Dense(self.n)
        self.norm1 = layers.BatchNormalization()
        self.act1 = layers.Activation(tf.nn.leaky_relu)

        self.dense2 = layers.Dense(self.n)
        self.norm2 = layers.BatchNormalization()
        self.act2 = layers.Activation(tf.nn.leaky_relu)

        self.dense3 = layers.Dense(self.o)
        self.norm3 = layers.BatchNormalization()
        self.act3 = layers.Activation(tf.nn.sigmoid)
        
    def call(self, inp, TRAINING=True):
        
        inp = self.act1(self.norm1(self.dense1(inp), training=TRAINING))
        inp = self.act2(self.norm2(self.dense2(inp), training=TRAINING))
        inp = self.act3(self.norm3(self.dense3(inp), training=TRAINING))
        
        return inp

class Discrinimator(Model):
    def __init__(self, num_nodes, logit_size):
        super(Discrinimator, self).__init__()
        self.n = num_nodes
        self.l = logit_size
                
        self.dense1 = layers.Dense(self.n)
        self.norm1 = layers.BatchNormalization()
        self.act1 = layers.Activation(tf.nn.leaky_relu)

        self.dense2 = layers.Dense(self.n)
        self.norm2 = layers.BatchNormalization()
        self.act2 = layers.Activation(tf.nn.leaky_relu)

        self.dense3 = layers.Dense(self.l)
        self.norm3 = layers.BatchNormalization()
        self.act3 = layers.Activation(tf.nn.sigmoid)
        
    def call(self, inp, TRAINING=True):
        
        inp = self.act1(self.norm1(self.dense1(inp), training=TRAINING))
        inp = self.act2(self.norm2(self.dense2(inp), training=TRAINING))
        inp = self.act3(self.norm3(self.dense3(inp), training=TRAINING))
        
        return inp

generator = Generator(256, dim)
discriminator = Discrinimator(256, dim)

generator(data_x)

def gain (data_x, gain_parameters):
  
    ## GAIN structure
    # Generator
    G_sample = generator(X, M)

    # Combine with observed data
    Hat_X = X * M + G_sample * (1-M)

    # Discriminator
    D_prob = discriminator(Hat_X, H)

    ## GAIN loss
    D_loss_temp = -tf.reduce_mean(M * tf.log(D_prob + 1e-8) \
                                + (1-M) * tf.log(1. - D_prob + 1e-8)) 

    G_loss_temp = -tf.reduce_mean((1-M) * tf.log(D_prob + 1e-8))

    MSE_loss = \
    tf.reduce_mean((M * X - M * G_sample)**2) / tf.reduce_mean(M)

    D_loss = D_loss_temp
    G_loss = G_loss_temp + alpha * MSE_loss 

    ## GAIN solver
    D_solver = tf.train.AdamOptimizer().minimize(D_loss, var_list=theta_D)
    G_solver = tf.train.AdamOptimizer().minimize(G_loss, var_list=theta_G)

    ## Iterations
    sess = tf.Session()
    sess.run(tf.global_variables_initializer())
   
    # Start Iterations
    for it in tqdm(range(iterations)):    
      
        # Sample batch
        batch_idx = sample_batch_index(no, batch_size)
        X_mb = norm_data_x[batch_idx, :]  
        M_mb = data_m[batch_idx, :]  
        # Sample random vectors  
        Z_mb = uniform_sampler(0, 0.01, batch_size, dim) 
        # Sample hint vectors
        H_mb_temp = binary_sampler(hint_rate, batch_size, dim)
        H_mb = M_mb * H_mb_temp

        # Combine random vectors with observed vectors
        X_mb = M_mb * X_mb + (1-M_mb) * Z_mb 

        _, D_loss_curr = sess.run([D_solver, D_loss_temp], 
                                  feed_dict = {M: M_mb, X: X_mb, H: H_mb})
        _, G_loss_curr, MSE_loss_curr = \
        sess.run([G_solver, G_loss_temp, MSE_loss],
                 feed_dict = {X: X_mb, M: M_mb, H: H_mb})
            
    ## Return imputed data      
    Z_mb = uniform_sampler(0, 0.01, no, dim) 
    M_mb = data_m
    X_mb = norm_data_x          
    X_mb = M_mb * X_mb + (1-M_mb) * Z_mb 

    imputed_data = sess.run([G_sample], feed_dict = {X: X_mb, M: M_mb})[0]

    imputed_data = data_m * norm_data_x + (1-data_m) * imputed_data

    # Renormalization
    imputed_data = renormalization(imputed_data, norm_parameters)  

    # Rounding
    imputed_data = rounding(imputed_data, data_x)  

    return imputed_data
