==> legacy/3_pg_input_processing_.ipynb <==
import os

import numpy as np
import pandas as pd

# Plant growth

### 2020 Summer

output_df = pd.read_csv('./results/2020_S/pg_reg_result.csv', index_col='Unnamed: 0')
output_label_2020S = output_df.values
data_indices_2020S = output_df.index

### 2020 Winter

output_df = pd.read_csv('./results/2020_W/pg_reg_result.csv', index_col='Unnamed: 0')
output_label_2020W = output_df.values
data_indices_2020W = output_df.index

OUTPUT_MAXS = np.concatenate([output_label_2020S, output_label_2020W], axis=0).max(axis=0)
OUTPUT_MINS = np.concatenate([output_label_2020S, output_label_2020W], axis=0).min(axis=0)

output_label_2020S = (output_label_2020S - OUTPUT_MINS)/(OUTPUT_MAXS - OUTPUT_MINS)
output_label_2020W = (output_label_2020W - OUTPUT_MINS)/(OUTPUT_MAXS - OUTPUT_MINS)

f = open('./results/pg_dataset.npz', 'wb')
np.savez(f,
         data_indices_2020S = data_indices_2020S,
         data_indices_2020W = data_indices_2020W,
         output_label_2020S = output_label_2020S,
         output_label_2020W = output_label_2020W,
         OUTPUT_MAXS = OUTPUT_MAXS,
         OUTPUT_MINS = OUTPUT_MINS
        )
f.close()

print(output_label_2020S.shape)
print(data_indices_2020S.shape)
print(output_label_2020W.shape)
print(data_indices_2020W.shape)
